feat(api) : refactored services
