+++ packages/api/src/controllers/deployment-connection.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get, Param, Post, Body, Put } from "@nestjs/common";
 import { CreateApikeyDto, DeploymentConnectionDTO, UpdateApikeyDto, UpdateDeploymentConnectionDTO } from "../core/dtos";
-import { DeploymentConnectionUseCases } from "../use-cases/deploymentConnection/deployment-connection.use-case";
+import { DeploymentConnectionUseCases } from "../use-cases/deployment-connection/deployment-connection.use-case";
 
 @Controller("deployment-connection")
 export class DeploymentConnectionController {
