+++ packages/api/src/controllers/index.ts
@@ -1,5 +1,5 @@
 export * from "./app.controller";
 export * from "./application.controller";
-export * from "./deploymentRecord.controller";
+export * from "./deployment-record.controller";
 export * from "./apikey.controller";
-export * from "./deploymentConnection.controller";
+export * from "./deployment-connection.controller";
