+++ packages/api/src/core/dtos/index.ts
@@ -1,6 +1,6 @@
 export * from "./application.dto";
-export * from "./deploymenetRecord.dto";
+export * from "./deploymenet-record.dto";
 export * from "./apikey.dto";
-export * from "./deploymentConnection.dto";
+export * from "./deployment-connection.dto";
 export * from "./property.dto";
 export * from "./environment.dto";
