+++ packages/api/src/core/entities/index.ts
@@ -1,8 +1,8 @@
-export * from "./deploymentRecord.entity";
+export * from "./deployment-record.entity";
 export * from "./application.entity";
 export * from "./apikey.entity";
-export * from "./deploymentConnection.entity";
+export * from "./deployment-connection.entity";
 export * from "./environment.entity";
 export * from "./event.entity";
-export * from "./eventTimeTrace.entity";
+export * from "./event-time-trace.entity";
 export * from "./property.entity";
