+++ packages/api/src/repository/mongo/model/index.ts
@@ -1,8 +1,8 @@
-export * from "./deploymentRecord.model";
+export * from "./deployment-record.model";
 export * from "./application.model";
 export * from "./apikey.model";
-export * from "./deploymentConnection.model";
+export * from "./deployment-connection.model";
 export * from "./environment.model";
 export * from "./event.model";
-export * from "./eventTimeTrace.model";
+export * from "./event-time-trace.model";
 export * from "./property.model";
